rename all folders to use kebab case to avoid git renaming errors in the future
