+++ src/app-common/location-detail/location-detail.css
