+++ src/app-pages/map/components/map-details/mapDetails.css
